FedBERT-LoRA/requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Federated Learning
flwr>=1.5.0

# Parameter Efficiency
peft>=0.4.0

# Utilities
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
